requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
openai==1.3.7
langchain==0.3.14
langchain-core==0.3.29
langchain-community==0.3.14
langchain-openai==0.2.14
langchain_huggingface-0.1.2
dashscope==1.20.14
python-multipart==0.0.20
tiktoken==0.5.2
sentence-transformers==3.3.1
docx2txt==0.8
elasticsearch==8.17.0
